Fix name display
Xogar was showing a lot of blanks and appid_ games. On the default
window size, it was confusing and when I made appid_ names return
empty strings, it looked like there were no games listed until I
scrolled down.

Instead, I made it to also check for other listings of the name in
the manifest file and if that still doesn't return anything then
we pick the installDir which is usually named after the game.

Anything still blank isn't listed because if it doesn't have an
installDir then it was probably uninstalled and the app is still
listed.